aplicatie realizata

+++ src/main/java/com/Bianca/SmartHouse/demo/HouseRepository.java
@@ -2,6 +2,6 @@ package com.Bianca.SmartHouse.demo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HouseRepository extends JpaRepository<House, Integer> {
+public interface HouseRepository extends JpaRepository<House, Long> {
 
 }
